Rename EuclidConstructions.py to PrebuiltConstructions.py to reflect recent refactor
